speed up code on C BLAS implementation
This was done by changing the way inputGradient() works. Before, it manually scaled and added each row to the result. This was fine on native BLAS, where there was very little overhead for BLAS calls. However, cgo takes some overhead to invoke C functions, and this overhead seems to have been very costly in this case: the LSTM benchmark used to take 9ms, now it takes 4ms with C BLAS.