added public func testing package
This will be useful for other projects which create Func and RFunc objects.

You might be wondering why I duplicated the func_test.go file rather than using the new functest package in the autofunc tests. At first, this might seem like a horrible violation of DRY. While this is true, there was not much else I could do. The functest package references autofunc, so autofunc cannot reference functest back in it's _test.go files (or else it's a circular import). This could be solved by moving all the _test.go files to functest as well, having all the tests for autofunc be in the functest package. However, I did not like this idea, because it means that the package's tests are no longer trivial to run.