fixed func tests for outputs that depend on Variables
A Result/RResult cannot be counted on to remain valid after its underlying variables or functions are modified. Up until now, none of my Result types actually depended on their underlying variables, because their Output() slices were allocated separately. However, assuming this will always be the case is false: it is perfectly fine for a Result/RResult's Output() or ROutput() to change or become nil when an underlying variable is changed (allowing for lazy evaluation, etc.).