remove VectorCache system
The cache system did not improve performance nearly as much as I had hoped. In real neural networks, so much of the overhead comes from matrix multiplication that nothing else really matters. GC is the least of one's worries when implementing a multithreaded neural net.

The only reason I thought GC mattered was that Go's profiler was showing weird results in multithreaded neuralnets. It turned out that the problem was CPU cache related, and memory was becoming a bottleneck (not allocation, but raw memory access). To fix this, I implemented batching. Before doing batching (i.e. before discovering the memory access issues), I implemented the VectorCache mechanism only to find that it was not beneficial.

I figured it was time to remove the VectorCache system, since it adds pointless clutter. Even if some pathological benchmarks are slower without the cache, it is not worth the extra complexity and work to maintain. Every single Func/RFunc needed to manage its memory using a cache, and other libraries (e.g. a neuralnet library) also had to maintain caches themselves. This resulted in completely unneeded bloat for such a marginal gain in performance. #ILoveGC.